Creation Time Pattern Example
I've had a few people be confused on the -creationtimestamp example for the folder path. They were actually  putting in folder instead of changing it to what their folder was thinking it was some kind of key word. I think by changing the example to something a little let generic it will help.